[catalog] logging for HMS-related operations
This patch adds more logging for results of HMS catalog's operations.
Since the number of DDL operations is usually not very high, this new
logging should not contribute significantly to the amount of information
logged.

The motivation for this patch is to increase the supportability of Kudu
clusters where HMS integration is enabled.

Change-Id: Iab09e9cdd9a994b62f728cdc218e5e905488a420
Reviewed-on: http://gerrit.cloudera.org:8080/16722
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>