Flexi Raft will considering current config to determine voter list and other changes.
Summary:
In Flexi Raft we use "Majority" commit rule along with a voter_distribution provided
to Raft by automation. The voter_distribution and the Majority rule together determine
the number of votes Flexi-Raft will wait for commit in region. So for a region with 1 database + 2 witnesses,
voter_distribution will say {"region1" : 3} and the Majority will mean that LEADER will wait for 2 ACKs.
However if we add 2 more members due to manual mitigations or during  replacements, we need to account
for the increased voter count. Now voter count = 5 and majority should be 3. This was not happening.

The fix is to consider #ACKS_REQUIRED = Majority(MAX(Current Voters, Voter Distribution)).
The voter distribution of 3 helps us in preventing #ACKS_REQUIRED to ever go below 2.
For new region turnups, we want to make sure that the voter distribution map for Raft.

We will disallow any Leader elections in regions where the voter distribution map is not present.
We will also disallow addition of voters without a proper voter distribution map.
This will check that automation like add-member is doing the right thing.
Flexi-Raft for Single Region mode currently operates in 2 ways (controlled by a flag)
1. We elect new LEADER by getting majority from current leader region.
2. We elect new LEADER by getting votes from majority of regions (majority of majority of regions)

Both these approaches had a gap where the majority in the Candidates region was not considered/mandated.
This can cause an availability issue where a single healthy voter in an unhealthy region can
win the election but wouldn't be able to establish authority on the Ring as it would
fail to push the No-Op (the other voters in that region are unhealthy or not present).
This change alters this behavior by waiting for majority from candidate
region.

Test Plan: Unit test to demonstrate the gaps and this patch will fix it,
is to be written on mysql raft side.

Reviewers: iRitwik, bhatvinay, yashtc, yichenshen

Subscribers:

Tasks:

Tags: