client-test: start txn on participant before inserting
fd7a40367a81555230eac62a1a8d1467433ed100 added pre-checks for
transactional inserts that the transaction exists on participants, but
was merged without rebasing to run ClientTest.TxnToken.

This starts the transaction on the participants before inserting as a
stopgap until we implement doing so automatically upon inserting for the
first time.

Change-Id: Ia513d90aadac4555d47174c91e4c51b4e158f350
Reviewed-on: http://gerrit.cloudera.org:8080/16757
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins