[mini-cluster] improved WaitForTabletsRunning
This patch adds an extra output parameter for the
ExternalMiniCluster::WaitForTabletsRunning() method to retrieve
information on currently running tablets.  This helps in test scenarios
where it's necessary to get the set of identifiers for running tablets
in the very beginning, avoiding an extra call to itest::ListTablets()
and building the TServerDetail map.

The newly added functionality is to be used in follow-up changelists.

Change-Id: I076064233c10f76b9c5c795d7943b10ef430ee90
Reviewed-on: http://gerrit.cloudera.org:8080/16764
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>