[CLI] introduce connection negotiation timeout
Since the connection negotiation timeout for Kudu client is now
configurable (see KUDU-2966 and changelist 62ec1e474), it makes sense
to allow using the new functionality in the kudu CLI tools.

This changelist introduces a new command line option/flag
--negotiation_timeout_ms to control the client-side connection
negotiation timeout in the kudu CLI tools.  The default value for the
new flag is set to 3000 milliseconds to match the default connection
negotiation timeout on the server.  With larger and heavy loaded
clusters it might be a good idea to the negotiation timeout if
connection negotiation timeouts happen time to time.  Also, it could
help in cases where the RPC authentication is required and KDC is not
fast enough.  Keep in mind that the total RPC timeout includes the
connection negotiation time, so in general it makes sense to bump
--timeout_ms along with --negotiation_timeout_ms by the same delta.

In addition, with this changeliest, all RPC-related kudu CLI tools
except for `kudu local_replica copy_from_remote` now have command line
arguments to control the RPC and the connection negotiation timeouts:
  * timeout_ms
  * negotiation_timeout_ms

In addition, this patch also unifies the way how KuduClient instances
are created in the kudu CLI code.

I'm planning to take care of the RPC-related parameters of the
`kudu local_replica copy_from_remote` tool in a separate patch.

Change-Id: I5627517fd1da4f4d7472f1f0af810d741efe3390
Reviewed-on: http://gerrit.cloudera.org:8080/16770
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>