[build] Support and require C++17
This patch adjust the build to start supporting and requiring C++17
(therefore GCC 7+). It adjusts the documentation and supported
OS matrix to match the new requirements as well.

I needed to replace std::random_shuffle with std::shuffle.
In C++14 `std::random_shuffle` is deprectated and in C++17 it is
removed all together. This patch replaces `std::random_shuffle` with
`std::shuffle` passing in a generator.

Additionally I needed to patch libev due to a small C++17 incompatibility:
https://github.com/enki/libev/pull/8

Change-Id: I15312473bda8c743eac228cd23c337fae5acf99d
Reviewed-on: http://gerrit.cloudera.org:8080/16776
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>