[tools] add missing top-level entries into TestModeHelp
This patch adds missing entries into the ToolTest.TestModeHelp scenario
for the existing top-level commands of the kudu CLI tool, extending
the coverage of the scenario to all currently existing top-level
kudu CLI commands.

I also did a minor clean-up on the code: rearranged the entries in the
scenario's code to match the sequence of those in the output of the
tool when run with no arguments or with --help flag.

Change-Id: I04c6744429e7a6a87ed6e87c1f5cab9c67d0f327
Reviewed-on: http://gerrit.cloudera.org:8080/16790
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>