client: allow tablet ID lookups from the MetaCache
Having a client interface for getting a RemoteTablet asynchronously will
be useful when it comes time to send participant ops to specific tablet
IDs from the TxnSystemClient.

In addition to the existing key-indexed cache entries, this patch
extends the MetaCache to track a separate set of cache entries by tablet
ID. There is no relationship between the two sets of cache entries -- a
tablet may exist in one but not the other.

The new id-based cache entries are only used if performing an id-based
lookup in the MetaCache via its new internal LookupTabletById() call
that will be used in a later patch.

Change-Id: Ib2333add5c3ab8403c48e69d29c90f3aec0914b6
Reviewed-on: http://gerrit.cloudera.org:8080/16794
Tested-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>