[build-support] devtoolset-8 allows to build on CentOS 6
RHEL/CentOS 6 is beyond its EOL date already, but it doesn't hurt
to keep Kudu building on old build servers (even if that hardware
is close to its own EOL as well).

With this patch, I was able to rebuild all thirdparty components
and Kudu itself in DEBUG and RELEASE configuration after installing
devtoolset-8 on old CentOS 6 build server.

Change-Id: I15f58edd090243ceb07a46f92d44e9998747ec5e
Reviewed-on: http://gerrit.cloudera.org:8080/16806
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>