[txn] update not-a-leader retry logic in TxnSystemClient
This patch updates the retry logic in TxnSystemClient to handle
the case when TxnStatusManager returns Status::ServiceUnavailable().

In addition, this patch adds extra provision to check for tablet replica
status in StatusManager::CheckTxnStatusDataLoadedUnlocked().
The signatures of the involved methods have been updated as well.

This patch also enhances FinalizeCommitTransaction() to report
on errors from underlying tablet, if any.  Also, TXN_ILLEGAL_STATE error
code is added into CoordinateTransactionResponsePB responses when
TxnStatusManager finds the requested transaction in a wrong state.

I didn't add new tests scenarios in this patch: they are present in
follow-up patches which introduce the tracking of stalled transactions.

Change-Id: I1b8abb27c7678c5c616a325343620902f6cbfd59
Reviewed-on: http://gerrit.cloudera.org:8080/16815
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>