Replace boost::iequals with our own implementation
This patch replaces uses of boost::iequals with our own implementation.
This is a step in our ongoing work to reduce dependence on boost.

Note that we don’t need to worry about locale awareness or performance
given all usage today is against ASCII strings (not locale specific) and are
not in performance sensitive code paths.

Change-Id: I6d2fdd30b739cee53e1551f784dbebfbe76e9233
Reviewed-on: http://gerrit.cloudera.org:8080/16826
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Grant Henke <granthenke@apache.org>