Snooze failure detection for a randomized timeout
Summary: This is to be avoid collisions during voting and being more
in line with Raft Thesis and LogCabin implementation. When a FOLLOWER
gets a UpdateReplica messge it snoozes its failure detector for a
randomized period from 1X - 2X of MinimumElectionTimeout. No change in
when it allows votes. That is still MinElectionTimeout. This prevents
wasted voting cycles and also increases the spread for voters so that
they don't collide with other candidates.

Also improved logging to have better messages distinguishing
pre-election/election and adding hostname_port to denied messages.

Added a new API to raft consensus log which allows MySQL Raft to
call the log abstraction with TruncateOpsAfter.

Test Plan: The testing of this was done as part of the RPM based
testing flow for Mysql Raft.

Reviewers:

Subscribers:

Tasks:

Tags: