Expose kudu pbc dump tool as library (#88)
Summary: This patch makes the kudu dump command, a part of the kudu CLI
tool, build again while keeping all other actions commented out of the
build script. It also exposes the CLI tool as a library instead of an
executable which makes integration with the downstream toolchain easier.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: