Improve preadv/pwritev with Xcode 12 and the OSX 11 SDK
This is a follow on to 0e2c288 which introduced version checks to handle
the introduction of preadv/pwritev in the OSX 11 SDK.

Instead of using a compile time macro to disable using the
preadv/pwritev simulations, this patch reverts to using them all the time
and adjusts the naming of the method to ensure there are no symbol
issues. This is important so that a binary compiled using the OSX 11 SDK
can run on a machine running OSX < 11. This scenario will frequently
occur on Xcode 12 given the OSX 11 SDK is installed by default and
chosen by default by CMake even though the machine is running OSX 10.

Change-Id: I90ba8f65b756c77c25dc776d0174b0d7b5678137
Reviewed-on: http://gerrit.cloudera.org:8080/16852
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>