[client] small cleanup on AsyncRandomTxnManagerRpc
This patch changes AsyncRandomTxnManagerRpc so it owns the request
protobuf message.  The rationale: there is no need to keep a pointer
because the request message is no longer needed at the call sites
once it's populated and passed to the AsyncRandomTxnManagerRpc's
constructor.  I also modified the call sites accordingly to use
move semantics for the corresponding argument while calling the
constructor.

This patch does not contain any functional changes.

Change-Id: Ide5b96644832bbcd2c3b08e72aee49f993ddb95f
Reviewed-on: http://gerrit.cloudera.org:8080/16867
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>