client: retry id-based lookups if there is no leader
In testing out a later patch, I stumbled on the fact that we may end
up never retrying a tablet lookup in cases where all cached replicas are
marked as followers. The correct behavior is to attempt another lookup
in hopes that the master has more up-to-date information about
leadership than the client.

This is only an issue with the new id-based MetaCache lookups. This
patch extends the id-based path with logic that was already in the
key-based lookup path that prevented this.

Along the way, I did some renaming for clarity.

Change-Id: If311f7e94a4fc2efa067476a4cef9ef5b378d981
Reviewed-on: http://gerrit.cloudera.org:8080/16875
Tested-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>