[thirdparty] Fix curl build on SLES
SLES does not put the krb5-config in /usr/bin as curl expects.
Instead krb5-config is installed into `/usr/lib/mit/bin/`.
Additionally it doesn’t put that location on the `PATH` by default,
which means that the `which` command can fail to find it.
Building curl with an empty KRB5CONFIG results in the following
error:
   configure: error: one or more libs available at link-time are not available run-time. Libs used at link-time: -lgssapi

This patch fixes the issue for more common/default environments by
falling back to the special SLES krb5_config location if one is not found
and the file exists.

Change-Id: I0961865a00601431bbc833360590cafa3aaf65c6
Reviewed-on: http://gerrit.cloudera.org:8080/16883
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins