[docker] Add support for openSUSE in the Docker build
This patch adds support for openSUSE/leap:15 as a base image in the
Docker build. This is useful for testing and validating Kudu usage on
openSUSE/SLES.

I fixed the logic in `get_os_tag` to handle the illegal `/` character
in the resulting tag. This results in a cleaner tag such as
`apache/kudu:latest-leap15`.

Change-Id: I9c1ce282fda67db5043445a139c888bc74c44680
Reviewed-on: http://gerrit.cloudera.org:8080/16885
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>