[consensus] Allow sending status-only request messages to FAILED peer
This change adds the ability for a leader to send status-only
messages to a peer even if it's in FAILED_UNRECOVERABLE state.
This ability is turned off by default and controlled via
a PeerMessageQueue parameter.

Without this change when the system catalog is copied externally
the new master remains in FAILED_UNRECOVERABLE state and doesn't get
promoted to being a VOTER despite the system catalog being up to date.
The procedure for end-to-end testing that hooks up masters to use
this Raft config is a separate change.

Change-Id: I229cc739c1b5ec7b11ce05d5e6b1b8e9d654d6f7
Reviewed-on: http://gerrit.cloudera.org:8080/16899
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>