[docs] Add note about Javadoc compatibility
An earlier commit (3d344ecefe) disabled building javadoc on JDK 10+ due
to some compatibility issues. When attempting to build javadoc, the
actual building will be skipped and it will fail making the docs.

If the javadoc build is re-enalbed for JDK 10+, it fails with the below
error:

javadoc: error - Class org.apache.yetus.audience.tools.IncludePublicAnnotationsStandardDoclet is not a valid doclet.
Note: As of JDK 13, the com.sun.javadoc API is no longer supported.

This commit adds a note about this compatibility issue to the docs
building readme to make sure people build docs on JDK 8.

Change-Id: I1a84edd44e890c37fcf857f7904b6de10948f0fb
Reviewed-on: http://gerrit.cloudera.org:8080/16924
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>