[FR] [perf] [1/n] perf optimization in watermark calulation
Summary:
AdvanceMajorityReplicatedWatermarkFlexiRaft() is a function that advances
queue watermarks in FlexiRaft. It does a bunch of expensive operations like:
(1) Compute a map of vectors (keyed by region name) with regional watermarks
(i.e a list of indexes for all peers indicating the last OpId that was acked by
that peer)
(2) Sort each of the vector in this map
(3) Build a voter_distribution map - i.e a map of number of voters in each
region (keyed by region name)
(4) Build another map with number of voters in each region based on
active_config
(5) Fix up the actual voters by iterating over maps from (3) and (4)

The key observation here is that in SINGLE_REGION_DYNAMIC mode, only the leader
region's voters can advance the commit index (i.e watermark). Hence, building
such maps is wasteful as it will be discarded immediately.

Fixes introduced in this diff:
(a) In SINGLE_REGION_DYNAMIC mode, instead of computing a map of vectors for
each region, a single vector containing last OpId for peers in leader region
is computed. This avoids map construction/destruction cost as well as additional
vector construction/destruction cost
(b) We need to sort only one vector instead of multiple vectors due to (1)
(c) No need to build another map with number of voters in each region. Just a
single integer with number of voters in leader region is computed. This also
eliminates the iteration and lookups in (5) mentioned above.
(d) Skip computing watermarks completely if the response being processed is
from a region other than leader region

Some minor part of the code had to be duplicated for efficiency and simplicity

We could also think of caching the voter_distribution map (and update it only
when a config change happens). Will explore this in a different PR.

Test Plan:
Existing tests for functional coverage.
Pending perf runs. Will update once I run those

Reviewers:

Subscribers:

Tasks:

Tags: