[java] Add JAAS example to client API doc
The AsyncKuduClient class documents how to connect to a secure Kudu
cluster using ticket cache or using a keytab when the application is
using hadoop-common libraries. We can't always rely on
UserGroupInformation though, so this commit expands the keytab-based
login case by a JAAS config example.

It also fixes a formatting issue with the "@Override" in the code, as it
wasn't properly rendered.

Change-Id: I7db5bf67727f7318272de7ea2156a33255a2dc8b
Reviewed-on: http://gerrit.cloudera.org:8080/16940
Tested-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>